Update project bootstrap to use standalone APIs
This is the result of running `ng g @angular/core:standalone` and choosing "Bootstrap the project using standalone APIs". All the checks are happy!

It's worth noting that there's no commit for the middle step (choosing "Remove unnecessary NgModule classes") because that step lead to no changes. Presumably we hadn't made any `NgModule` classes.
